refactor(frontend): reuse service account JSON uploader in org agent credentials
