+++ frontend/src/components/provider-config-form.tsx
@@ -1,23 +1,12 @@
 "use client"
 
 import { zodResolver } from "@hookform/resolvers/zod"
-import { File as FileIcon, Info, Save, Upload, X } from "lucide-react"
-import {
-  type ChangeEvent,
-  type DragEvent,
-  useCallback,
-  useMemo,
-  useRef,
-  useState,
-} from "react"
-import {
-  type ControllerRenderProps,
-  type UseFormReturn,
-  useForm,
-  useWatch,
-} from "react-hook-form"
+import { Info, Save } from "lucide-react"
+import { useCallback, useMemo } from "react"
+import { useForm, useWatch } from "react-hook-form"
 import { z } from "zod"
 import type { IntegrationUpdate, ProviderRead } from "@/client"
+import { ServiceAccountJsonUploader } from "@/components/service-account-json-uploader"
 import { MultiTagCommandInput } from "@/components/tags-input"
 import { Alert, AlertDescription } from "@/components/ui/alert"
 import { Badge } from "@/components/ui/badge"
@@ -34,7 +23,6 @@ import {
 import { Input } from "@/components/ui/input"
 import { useIntegrationProvider } from "@/lib/hooks"
 import { isMCPProvider } from "@/lib/providers"
-import { cn } from "@/lib/utils"
 import { useWorkspaceId } from "@/providers/workspace-id"
 
 type EndpointHelp = ProviderRead["authorization_endpoint_help"]
@@ -354,10 +342,37 @@ export function ProviderConfigForm({
                     <FormControl>
                       {isServiceAccountProvider ? (
                         <ServiceAccountJsonUploader
-                          field={field}
-                          form={form}
+                          value={field.value ?? ""}
+                          onChange={field.onChange}
+                          onError={(message) => {
+                            form.setError("client_secret", {
+                              type: "manual",
+                              message,
+                            })
+                          }}
+                          onClearError={() => {
+                            form.clearErrors("client_secret")
+                          }}
+                          onDetectedEmail={(email) => {
+                            const currentClientId = form
+                              .getValues("client_id")
+                              ?.trim()
+                            if (
+                              (!currentClientId ||
+                                currentClientId.length === 0) &&
+                              email
+                            ) {
+                              form.setValue("client_id", email, {
+                                shouldDirty: true,
+                                shouldTouch: true,
+                              })
+                            }
+                          }}
                           placeholder={clientSecretPlaceholder}
                           existingConfigured={hasExistingSecret}
+                          hasError={Boolean(
+                            form.formState.errors.client_secret
+                          )}
                         />
                       ) : (
                         <Input
@@ -553,221 +568,6 @@ export function ProviderConfigForm({
         </form>
       </Form>
     </div>
-  )
-}
-
-interface ServiceAccountJsonUploaderProps {
-  field: ControllerRenderProps<OAuthSchema, "client_secret">
-  form: UseFormReturn<OAuthSchema>
-  placeholder: string
-  existingConfigured: boolean
-}
-
-function ServiceAccountJsonUploader({
-  field,
-  form,
-  placeholder,
-  existingConfigured,
-}: ServiceAccountJsonUploaderProps) {
-  const fileInputRef = useRef<HTMLInputElement | null>(null)
-  const [fileName, setFileName] = useState<string | null>(null)
-  const [detectedEmail, setDetectedEmail] = useState<string | null>(null)
-
-  const hasError = Boolean(form.formState.errors.client_secret)
-
-  const resetInput = useCallback(() => {
-    if (fileInputRef.current) {
-      fileInputRef.current.value = ""
-    }
-  }, [])
-
-  const clearSelection = useCallback(() => {
-    setFileName(null)
-    setDetectedEmail(null)
-    field.onChange("")
-    form.clearErrors("client_secret")
-    resetInput()
-  }, [field, form, resetInput])
-
-  const handleFile = useCallback(
-    (file: File | undefined) => {
-      if (!file) {
-        return
-      }
-
-      if (!file.name.toLowerCase().endsWith(".json")) {
-        field.onChange("")
-        setFileName(null)
-        setDetectedEmail(null)
-        form.setError("client_secret", {
-          type: "manual",
-          message: "Upload a .json file exported from Google Cloud.",
-        })
-        resetInput()
-        return
-      }
-
-      const reader = new FileReader()
-      reader.onload = () => {
-        const text = typeof reader.result === "string" ? reader.result : ""
-
-        try {
-          const parsed = JSON.parse(text)
-          if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) {
-            throw new Error("Uploaded key must be a JSON object.")
-          }
-          if (parsed.type !== "service_account") {
-            throw new Error('JSON key must include "type": "service_account".')
-          }
-          if (
-            typeof parsed.private_key !== "string" ||
-            parsed.private_key.trim().length === 0
-          ) {
-            throw new Error("JSON key is missing a private_key.")
-          }
-
-          const normalized = JSON.stringify(parsed)
-          field.onChange(normalized)
-          setFileName(file.name)
-
-          const email =
-            typeof parsed.client_email === "string"
-              ? parsed.client_email.trim()
-              : ""
-          setDetectedEmail(email || null)
-
-          const currentClientId = form.getValues("client_id")?.trim()
-          if ((!currentClientId || currentClientId.length === 0) && email) {
-            form.setValue("client_id", email, {
-              shouldDirty: true,
-              shouldTouch: true,
-            })
-          }
-
-          form.clearErrors("client_secret")
-        } catch (error) {
-          field.onChange("")
-          setFileName(null)
-          setDetectedEmail(null)
-          const message =
-            error instanceof Error
-              ? error.message
-              : "Failed to parse service account JSON key."
-          form.setError("client_secret", {
-            type: "manual",
-            message,
-          })
-        } finally {
-          resetInput()
-        }
-      }
-      reader.onerror = () => {
-        field.onChange("")
-        setFileName(null)
-        setDetectedEmail(null)
-        form.setError("client_secret", {
-          type: "manual",
-          message: "Failed to read the uploaded file.",
-        })
-        resetInput()
-      }
-
-      reader.readAsText(file)
-    },
-    [field, form, resetInput]
-  )
-
-  const handleInputChange = useCallback(
-    (event: ChangeEvent<HTMLInputElement>) => {
-      const file = event.target.files?.[0]
-      handleFile(file)
-    },
-    [handleFile]
-  )
-
-  const handleDrop = useCallback(
-    (event: DragEvent<HTMLDivElement>) => {
-      event.preventDefault()
-      const file = event.dataTransfer.files?.[0]
-      handleFile(file)
-      event.dataTransfer.clearData()
-    },
-    [handleFile]
-  )
-
-  const handleDragOver = useCallback((event: DragEvent<HTMLDivElement>) => {
-    event.preventDefault()
-    event.dataTransfer.dropEffect = "copy"
-  }, [])
-
-  return (
-    <>
-      <input
-        ref={fileInputRef}
-        type="file"
-        accept=".json,application/json"
-        className="hidden"
-        onChange={handleInputChange}
-      />
-      <div
-        onDrop={handleDrop}
-        onDragOver={handleDragOver}
-        className={cn(
-          "flex flex-col gap-3 rounded-md border border-dashed p-4 transition-colors",
-          hasError
-            ? "border-destructive/80 bg-destructive/5"
-            : "border-muted-foreground/30 bg-muted/40 hover:border-muted-foreground/50"
-        )}
-      >
-        <div className="flex flex-col gap-2">
-          <div className="flex flex-wrap items-center justify-between gap-3">
-            <p className="text-xs text-muted-foreground">
-              {fileName ? (
-                <span className="inline-flex items-center gap-1 text-sm font-medium text-foreground">
-                  <FileIcon className="h-4 w-4" />
-                  {fileName}
-                </span>
-              ) : (
-                placeholder
-              )}
-            </p>
-            <div className="flex items-center gap-2">
-              <Button
-                type="button"
-                variant="outline"
-                size="sm"
-                onClick={() => fileInputRef.current?.click()}
-              >
-                <Upload className="mr-2 h-4 w-4" />
-                {fileName ? "Replace file" : "Choose JSON"}
-              </Button>
-              {fileName && (
-                <Button
-                  type="button"
-                  variant="ghost"
-                  size="sm"
-                  onClick={clearSelection}
-                >
-                  <X className="mr-1 h-4 w-4" />
-                  Remove
-                </Button>
-              )}
-            </div>
-          </div>
-          {existingConfigured && !fileName && (
-            <p className="text-xs text-muted-foreground">
-              Existing key remains active until you provide a new file.
-            </p>
-          )}
-          {detectedEmail && (
-            <p className="text-xs text-muted-foreground">
-              Detected service account:{" "}
-              <span className="font-medium">{detectedEmail}</span>
-            </p>
-          )}
-        </div>
-      </div>
-    </>
   )
 }
